requirements.txt
celery==5.3.4
redis==5.0.1
docker>=7.0.0
flake8==6.1.0
radon==6.0.1
supabase==2.0.2
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
boto3==1.0.0
black==25.1.0